misc: separate gateway and scheduler
Link: https://code.alibaba-inc.com/cuikuilong.ckl/llumnix/codereview/25760449

+++ pkg/llm-gateway/service/handler/pd_vllm_mooncake_backend.go
@@ -35,7 +35,7 @@ func NewPdSplitVllmMoonCakeBackend(schMode types.ScheduleMode) (InferenceBackend
 
 func (b *PdSplitVllmMoonCakeBackend) buildPrefillRequestData(req *types.RequestContext) ([]byte, error) {
 	cmplReq := *(req.LLMRequest.CompletionRequest)
-	maxTokens := 1
+	maxTokens := uint64(1)
 	cmplReq.MaxTokens = &maxTokens
 	cmplReq.Stream = false
 	cmplReq.StreamOptions = nil
